feat: 이벤트 API 유저 인증 구현

+++ src/modules/events/event.service.ts
@@ -4,16 +4,20 @@ import { CreateEventRequest } from './dto/request/create.event.request';
 import { GetEventResponse } from './dto/response/get.event.response';
 import { EventEntity } from './entities/event.entity';
 import { GetEventListQueryRequest } from './dto/request/get.event.query.request';
+import { ForbiddenAccessException } from '../../global/exception/custom.exception';
 
 @Injectable()
 export class EventService {
     constructor(private readonly eventRepository: EventRepository) {}
 
     async createEvent(
+        userId: number,
         createEventRequest: CreateEventRequest,
     ): Promise<GetEventResponse> {
-        const eventEntity: EventEntity =
-            await this.eventRepository.createEvent(createEventRequest);
+        const eventEntity: EventEntity = await this.eventRepository.createEvent(
+            userId,
+            createEventRequest,
+        );
         return new GetEventResponse(eventEntity);
     }
 
@@ -32,17 +36,28 @@ export class EventService {
     }
 
     async updateEvent(
+        userId: number,
         eventId: number,
         updateEventRequest: CreateEventRequest,
     ): Promise<GetEventResponse> {
-        const event: EventEntity = await this.eventRepository.updateEvent(
-            eventId,
-            updateEventRequest,
-        );
-        return new GetEventResponse(event);
+        const event = await this.eventRepository.findById(eventId);
+
+        if (event.userId !== userId) {
+            throw new ForbiddenAccessException();
+        }
+
+        const updatedEvent: EventEntity =
+            await this.eventRepository.updateEvent(eventId, updateEventRequest);
+        return new GetEventResponse(updatedEvent);
     }
 
-    async deleteEvent(eventId: number): Promise<void> {
+    async deleteEvent(userId: number, eventId: number): Promise<void> {
+        const event = await this.eventRepository.findById(eventId);
+
+        if (event.userId !== userId) {
+            throw new ForbiddenAccessException();
+        }
+
         return this.eventRepository.deleteEvent(eventId);
     }
 }
